[wasm][debugger][tests] Extract Array tests to a new file
